Updated PUF styling to work with new carousel.js changes
Reworked a bunch of styling rules in puf.css to make the new carousel changes behave as needed.

Also added a new wrapper to the carousel cards in puf.js to simplify aligning cards with/without the highlight borders.

Still need to test the container height when loaded on mobile.
And have a lot of unused js script that can be removed in puf.js