Refactoring of puf block
Removed all slider scroll logic from puf.js, it now uses the logic from carousel.js.

Also removed all toggling of 1-col and 2-col layouts and the selected slide controls.

Replaced 4 resize observes with a single observer.
And added two intersection observers to handle adjusting the containers max height based on the selected slide.
I also fixed the sticky arrow control behaviour, and connected that to the intersection observers to also get the appropriate height.

Still a number of functions to review and refactor in puf.js
And a lot of classes can be cleaned up and improved in puf.css.