evaluate if breadcrumbs exist early and adjust the gnav height to reduce CLS
Added a function "hasBreadcrumbs" to scripts.js to evaluate if a page has breadcrumbs or not. And applies a new class "breadcrumbs-spacing" to the body if breadcrumbs exist.

In gnav.js the metadata tag for toggling breadcrumbs has been switched to Milos metadata tag in buildBreadCrumbArray function. And a couple lines in that function have been moved below a potential return condition to where they are used.

A new css variable was added for the breadcrumb height and a rule applying it with the new body class.