Lowered intersection threshold
Lowered the threshold that triggers intersection from 100% to 75%, and removed the debug alerts from the previous commit.

My suspicion is the addition of the faders/arrows/triggers is taking up some amount of width so that the card does not alway intersect it's parent container 100% on mobile.
This seems to be somewhat inconsistent, sometimes on refresh it was working for me, some times it wasn't. But I believe a lower threshold should correct the behaviour.