Moved initToggleTrigger outside of intersection observer
The line that initialized the toggle triggers has been moved outside of the callback for the carousel intersection observer.

This was preventing the toggles from being created until after that carousel was intersecting the viewport, which was creating issues on slower devices.

(to test throttle the CPU performance by 6x on the pricing page)