Milo Table block switch to custom block
Removed bulk of milo table logic in table-ax.js and built in all the logic needed for the the new feature table block for AX.

Removed a large portion of milo CSS from table-ax.css and built in the new custom styling. This file could still stand to be cleaned up quite a bit.
Left some comments in an effort to group CSS rules so other devs cold follow or clean up.
And some temporary llines to point to where to add/edit text styling rules, as I'm not sure if text should be styled in the block css or in the authored doc.

Also added 4 icons. Collapse and Expand icons in the icons folder, and the check and dash icons in the block's directory.