Got code into our repo (#577)
* Initial integration of the Milo Table block as is

Adding the untouched Milo table block code to the express project.
- Added the table js and css files named table-ax as requested by the AX team.
- Appended the Milo global styling to styles.css.
- Added the decorate.js file that tables.js needs.
- Added the MILO_EVENTS object to the utils.js file

This code is currently broken as import statements point to milo's file structure, and the block's styling is not applying as milo's styling targets .table not .table-ax.
(The AX team wanted an initial commit of the original Milo block)

* Repaired broken imports

Corrected all import paths relating to the milo table block to point to the proper express file locations.

(the block is no longer 'broken' but still not functional)

* Replaced block name in table.css

Replaced all milo styles in table.css that target .table to target the newly names .table-ax

* Overriding pricing styling from styles.css

Added a simple style rule to negate some questionably specific styling with a very general targeting in styles.css.

I don't imagine that style should be applied to any .pricing element in main...

* Milo Table block switch to custom block

Removed bulk of milo table logic in table-ax.js and built in all the logic needed for the the new feature table block for AX.

Removed a large portion of milo CSS from table-ax.css and built in the new custom styling. This file could still stand to be cleaned up quite a bit.
Left some comments in an effort to group CSS rules so other devs cold follow or clean up.
And some temporary llines to point to where to add/edit text styling rules, as I'm not sure if text should be styled in the block css or in the authored doc.

Also added 4 icons. Collapse and Expand icons in the icons folder, and the check and dash icons in the block's directory.

* Various improvements and fixes

Authored all of the one-plan pricing page feature table content to my test document, and had to tweak various aspects (table width for smaller column tables, parsing the headers to target {{dark}} buttons, etc.)

The content looks generally on par with the design doc at this stage.

* Fixed feature row header cols with no children not getting styled

Updated the handleSection function to wrap feature rows header column text in a p tag if it as not child elements.

* Fixed issues with tables not using spacer rows

Fixed issues with missing row borders when not using spacer rows.

---------

Co-authored-by: Aaron Atkinson <38330431+AaronAtkinson@users.noreply.github.com>